fixed showing other species in list

+++ grails-app/domain/species/Metadata.groovy
@@ -55,13 +55,13 @@ abstract class Metadata {
         }
     }
 
-    double getLatitude() {
+    float getLatitude() {
         if(topology)
-            return topology.getCentroid().getY();
+            return (float) topology.getCentroid().getY();
     }
 
-    double getLongitude() {
+    float getLongitude() {
         if(topology) 
-            return topology.getCentroid().getX();
+            return (float) topology.getCentroid().getX();
     }
 }
